refactor(tests): split wear tests into new file
